python/requirements.txt
psycopg2-binary
redis==5.0.7
python-dotenv==1.0.1